+++ services/QuillLMS/client/app/bundles/PremiumHub/containers/UsageSnapshotsContainer.tsx
@@ -1,16 +1,15 @@
 import React from 'react'
-import "react-dates/initialize";
-import moment from 'moment'
-import { DateRangePicker } from 'react-dates';
 
+import CustomDateModal from '../components/usage_snapshots/customDateModal'
 import SnapshotSection from '../components/usage_snapshots/snapshotSection'
+import Filters from '../components/usage_snapshots/filters'
 import { snapshotSections, TAB_NAMES, ALL, CUSTOM, } from '../components/usage_snapshots/shared'
-import { Spinner, DropdownInput, KEYDOWN, MOUSEDOWN } from '../../Shared/index'
+import { Spinner, DropdownInput, } from '../../Shared/index'
+import useWindowSize from '../../Shared/hooks/useWindowSize';
 
-const removeSearchTokenSrc = `${process.env.CDN_URL}/images/pages/administrator/remove_search_token.svg`
+const MAX_VIEW_WIDTH_FOR_MOBILE = 850
 
-const rightArrowImg = <img alt="" src={`${process.env.CDN_URL}/images/pages/administrator/right_arrow.svg`} />
-const leftArrowImg = <img alt="" src={`${process.env.CDN_URL}/images/pages/administrator/left_arrow.svg`} />
+const filterIconSrc = `${process.env.CDN_URL}/images/icons/icons-filter.svg`
 
 const filterData = {
   timeframes: [
@@ -48,54 +47,6 @@ const filterData = {
   ]
 }
 
-const CustomDateModal = ({ close, passedStartDate, setCustomDates, passedEndDate, }) => {
-  const [focusedInput, setFocusedInput] = React.useState("startDate")
-  const [startDate, setStartDate] = React.useState(passedStartDate)
-  const [endDate, setEndDate] = React.useState(passedEndDate)
-  const [buttonSectionStyle, setButtonSectionStyle] = React.useState({})
-
-  function handleDateChange(newDates) {
-    setStartDate(newDates.startDate)
-    setEndDate(newDates.endDate)
-  }
-
-  function handleClickApply() {
-    setCustomDates(startDate, endDate)
-  }
-
-  function handleSetFocusedInput(newFocusedInput) {
-    if (!newFocusedInput) { return }
-
-    setFocusedInput(newFocusedInput)
-  }
-
-  function isOutsideRange() { return false }
-
-  return (
-    <div className="modal-container custom-date-modal-container">
-      <div className="modal-background" />
-      <div className="custom-date-modal quill-modal modal-body">
-        <h2>Select custom date range</h2>
-        <DateRangePicker
-          autoFocus={true}
-          endDate={endDate}
-          focusedInput={focusedInput}
-          isOutsideRange={isOutsideRange}
-          navNext={rightArrowImg}
-          navPrev={leftArrowImg}
-          onDatesChange={handleDateChange}
-          onFocusChange={handleSetFocusedInput}
-          startDate={startDate}
-        />
-        <div className="button-section">
-          <button className="quill-button medium secondary outlined focus-on-light" onClick={close} type="button">Cancel</button>
-          <button className="quill-button medium primary contained focus-on-light" onClick={handleClickApply} type="button">Apply</button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
 const Tab = ({ section, setSelectedTab, selectedTab }) => {
   function handleSetSelectedTab() { setSelectedTab(section) }
 
@@ -105,17 +56,6 @@ const Tab = ({ section, setSelectedTab, selectedTab }) => {
   }
 
   return <button className={className} onClick={handleSetSelectedTab} type="button">{section}</button>
-}
-
-const SearchToken = ({ searchItem, onRemoveSearchItem, }) => {
-  function handleRemoveSearchItem() { onRemoveSearchItem(searchItem)}
-
-  return (
-    <div className="search-token">
-      <span>{searchItem.label}</span>
-      <button aria-label={`Remove ${searchItem.label} from filtered list`} className="interactive-wrapper focus-on-light" onClick={handleRemoveSearchItem} type="button"><img alt="" src={removeSearchTokenSrc} /></button>
-    </div>
-  )
 }
 
 const UsageSnapshotsContainer = ({}) => {
@@ -131,20 +71,15 @@ const UsageSnapshotsContainer = ({}) => {
   const [customEndDate, setCustomEndDate] = React.useState(null)
   const [searchCount, setSearchCount] = React.useState(0)
   const [selectedTab, setSelectedTab] = React.useState(ALL)
-  const [filterButtonsAreFixed, setFilterButtonsAreFixed] = React.useState(true)
   const [showCustomDateModal, setShowCustomDateModal] = React.useState(false)
   const [lastUsedTimeframe, setLastUsedTimeframe] = React.useState(null)
+  const [showMobileFilterMenu, setShowMobileFilterMenu] = React.useState(false)
+
+  const size = useWindowSize()
 
   React.useEffect(() => {
     getFilters()
   }, [])
-
-  React.useEffect(() => {
-    const el = document.getElementsByClassName('home-footer')[0]
-    const observer = new IntersectionObserver(([entry]) => { entry.isIntersecting ? setFilterButtonsAreFixed(false) : setFilterButtonsAreFixed(true); });
-
-    el && observer.observe(el);
-  }, []);
 
   React.useEffect(() => {
     if (selectedSchools !== allSchools || selectedGrades !== allGrades || selectedTimeframe !== defaultTimeframe(allTimeframes)) {
@@ -157,6 +92,10 @@ const UsageSnapshotsContainer = ({}) => {
 
     setSelectedTimeframe(lastUsedTimeframe)
   }, [showCustomDateModal])
+
+  function openMobileFilterMenu() { setShowMobileFilterMenu(true) }
+
+  function closeMobileFilterMenu() { setShowMobileFilterMenu(false) }
 
   function handleSetSelectedTimeframe(timeframe) {
     setLastUsedTimeframe(selectedTimeframe)
@@ -208,37 +147,13 @@ const UsageSnapshotsContainer = ({}) => {
     closeCustomDateModal()
   }
 
+  function handleSetSelectedTabFromDropdown(option) { setSelectedTab(option.value) }
+
   if (loadingFilters) {
     return <Spinner />
   }
 
   function handleClickDownloadReport() { window.print() }
-
-  function handleRemoveSchool(school) {
-    const newSchools = selectedSchools.filter(s => s.id !== school.id)
-    setSelectedSchools(newSchools)
-  }
-
-  function handleRemoveGrade(grade) {
-    const newGrades = selectedGrades.filter(g => g.value !== grade.value)
-    setSelectedGrades(newGrades)
-  }
-
-  const schoolSearchTokens = selectedSchools !== allSchools && selectedSchools.map(s => (
-    <SearchToken
-      key={s.id}
-      onRemoveSearchItem={handleRemoveSchool}
-      searchItem={s}
-    />
-  ))
-
-  const gradeSearchTokens = selectedGrades !== allGrades && selectedGrades.map(grade => (
-    <SearchToken
-      key={grade.value}
-      onRemoveSearchItem={handleRemoveGrade}
-      searchItem={grade}
-    />
-  ))
 
   const tabs = TAB_NAMES.map(s => (
     <Tab
@@ -248,6 +163,19 @@ const UsageSnapshotsContainer = ({}) => {
       setSelectedTab={setSelectedTab}
     />
   ))
+
+  const tabDropdownOptions = TAB_NAMES.map(tab => ({ value: tab, label: tab, }))
+
+  const tabDropdown = (
+    <DropdownInput
+      handleChange={handleSetSelectedTabFromDropdown}
+      id="tab-dropdown"
+      isSearchable={false}
+      label=""
+      options={tabDropdownOptions}
+      value={tabDropdownOptions.find(opt => opt.value === selectedTab)}
+    />
+  )
 
   const sectionsToShow = selectedTab === ALL ? snapshotSections : snapshotSections.filter(s => s.name === selectedTab)
   const snapshotSectionComponents = sectionsToShow.map(section => (
@@ -263,17 +191,21 @@ const UsageSnapshotsContainer = ({}) => {
     />
   ))
 
-  function renderFilterButtons(alwaysShow) {
-    if (!hasAdjustedFilters) { return null }
-
-    if (!alwaysShow && !filterButtonsAreFixed) { return null }
-
-    return (
-      <div className={`filter-buttons ${filterButtonsAreFixed && !alwaysShow ? 'fixed' : ''}`}>
-        <button className="quill-button small outlined secondary focus-on-light" onClick={clearFilters} type="button">Clear filters</button>
-        <button className="quill-button small contained primary focus-on-light" onClick={applyFilters} type="button">Apply filters</button>
-      </div>
-    )
+  const filterProps = {
+    allTimeframes,
+    allSchools,
+    allGrades,
+    applyFilters,
+    clearFilters,
+    selectedGrades,
+    setSelectedGrades,
+    hasAdjustedFilters,
+    handleSetSelectedTimeframe,
+    selectedTimeframe,
+    selectedSchools,
+    setSelectedSchools,
+    closeMobileFilterMenu,
+    showMobileFilterMenu,
   }
 
   return (
@@ -286,52 +218,22 @@ const UsageSnapshotsContainer = ({}) => {
           setCustomDates={setCustomDates}
         />
       )}
-      <section className="filter-container">
-        <div className="filters">
-          <label className="filter-label" htmlFor="timeframe-filter">Timeframe</label>
-          <DropdownInput
-            handleChange={handleSetSelectedTimeframe}
-            id="timeframe-filter"
-            isSearchable={false}
-            label=""
-            options={allTimeframes}
-            value={selectedTimeframe}
-          />
-          <label className="filter-label" htmlFor="school-filter">School</label>
-          <DropdownInput
-            handleChange={setSelectedSchools}
-            id="school-filter"
-            isMulti={true}
-            isSearchable={true}
-            label=""
-            options={allSchools}
-            optionType='school'
-            value={selectedSchools}
-          />
-          <div className="search-tokens">{schoolSearchTokens}</div>
-          <label className="filter-label" htmlFor="grade-filter">Grade</label>
-          <DropdownInput
-            handleChange={setSelectedGrades}
-            id="grade-filter"
-            isMulti={true}
-            isSearchable={true}
-            label=""
-            options={allGrades}
-            optionType='grade'
-            value={selectedGrades}
-          />
-          <div className="search-tokens">{gradeSearchTokens}</div>
-        </div>
-        {renderFilterButtons(true)}
-        {renderFilterButtons(false)}
-      </section>
+      <Filters
+        {...filterProps}
+      />
       <main>
         <div className="header">
           <h1>Usage Snapshot Report</h1>
           <button className="quill-button contained primary medium focus-on-light" onClick={handleClickDownloadReport} type="button">Download Report</button>
         </div>
         <div aria-hidden={true} className="tabs">
-          {tabs}
+          {size.width >= MAX_VIEW_WIDTH_FOR_MOBILE ? tabs : tabDropdown}
+        </div>
+        <div className="filter-button-container">
+          <button className="interactive-wrapper focus-on-light" onClick={openMobileFilterMenu} type="button">
+            <img alt="Filter icon" src={filterIconSrc} />
+            Filters
+          </button>
         </div>
         <div className="sections">
           {snapshotSectionComponents}
